Finish stake NonNegI64 chain; make split_byte seed truncation checked
Reject negative stakes in bonds_parser and carry NonNegI64 through
contracts.Validator.stake, build_bonds_map, node_launch and
transaction_balances. Replace the silent `i as u8`/`(len+i) as u8`/
`i as u16` seed truncation with checked `u8::try_from`/`u16::try_from`
at the replay/proposer/reduce sites, so an oversized deploy list is an
error rather than a silent wrap of the block/replay seed.

+++ node/src/revvaultexport/transaction_balances.rs
@@ -180,7 +180,7 @@ pub fn generate_rev_account_from_wallet_and_bond(
             .get(&pos_addr)
             .cloned()
             .unwrap_or_else(|| pos_vault.clone());
-        account_map.insert(pos_addr.clone(), current.receive_rev(*bond_amount));
+        account_map.insert(pos_addr.clone(), current.receive_rev(i64::from(*bond_amount)));
     }
     Ok(account_map)
 }
